refactor: validation 클래스 삭제
- 유효성 검증을 따로 나누니까 Input, Lotto 등 클래스의 역할이 약해지는 것 같아 삭제
